refactor pdf strategies into strategies
